refactor(macOS): consolidate macOS runtime code into src/os_utils/mac.py
Fold the per-feature darwin modules (permissions, notifications, panic tray,
config-window CGEventTap) and keyboard listener backends into os_utils/mac.py so
it is the single macOS platform backend. Move macOS-only tests under macos/ and
drop the intermediate macos/*.py package.
