+++ edgeware/macos/edgeware.spec
@@ -97,6 +97,9 @@ a = Analysis(
         'voluptuous',
         'tkinterweb', 'tkinter_tooltip',
         'AppKit', 'Foundation', 'Quartz',
+        'os_utils', 'os_utils.mac',
+        'os_utils.keyboard_listener_darwin',
+        'os_utils.keyboard_subprocess',
     ],
     hookspath=[],
     hooksconfig={},
