+++ edgeware/src/features/misc.py
@@ -31,167 +31,15 @@ import pystray
 from config.settings import Settings
 from desktop_notifier.common import Attachment, Icon
 from desktop_notifier.sync import DesktopNotifierSync
-from os_utils import make_shortcut, set_wallpaper
+from os_utils import check_accessibility_permission, check_input_monitoring_permission, make_shortcut, send_panic_tray, set_wallpaper, show_accessibility_dialog, show_input_monitoring_dialog
 from pack import Pack
-from panic import panic, send_panic_tray
+from panic import panic
 from paths import CustomAssets, Process
 from PIL import Image
 from pynput import keyboard
 from pypresence import Presence
 from roll import roll
 from state import State
-
-
-def check_accessibility_permission() -> bool:
-    """Check if the app has Accessibility permission on macOS.
-    
-    Returns True if permission is granted or if not on macOS.
-    Returns False if permission is not granted or if check fails.
-    """
-    if sys.platform != "darwin":
-        return True
-    
-    try:
-        import ctypes
-        import ctypes.util
-        
-        # Load ApplicationServices framework
-        framework_path = ctypes.util.find_library('ApplicationServices')
-        if not framework_path:
-            logging.warning("Could not find ApplicationServices framework")
-            return False
-        
-        app_services = ctypes.cdll.LoadLibrary(framework_path)
-        
-        # AXIsProcessTrusted() -> Boolean
-        app_services.AXIsProcessTrusted.restype = ctypes.c_bool
-        trusted = app_services.AXIsProcessTrusted()
-        
-        logging.info(f"Accessibility permission check: {'granted' if trusted else 'not granted'}")
-        return trusted
-    except Exception as e:
-        logging.warning(f"Could not check Accessibility permission: {e}")
-        return False  # Assume not granted if check fails
-
-
-def show_accessibility_dialog(root: Tk) -> None:
-    """Show dialog explaining Accessibility permission requirement.
-
-    If the user opts to open System Preferences, warn that a restart is required
-    for the change to take effect, then quit the app so they can relaunch it.
-    If they decline, keep running (the tray / IPC panic path still works).
-    """
-    from tkinter import messagebox
-
-    message = (
-        "Edgeware++ needs Accessibility permission to listen for the global panic hotkey.\n\n"
-        "To grant permission:\n"
-        "1. Open System Preferences\n"
-        "2. Go to Privacy & Security > Accessibility\n"
-        "3. Click the lock icon and enter your password\n"
-        "4. Check the box next to Edgeware++\n\n"
-        "If you don't want to grant this permission, you can still use:\n"
-        "• The Panic.app (separate application)\n"
-        "• The tray icon menu\n"
-        "• The Legacy Panic Key (requires focusing on a popup)\n\n"
-        "Do you want to open System Preferences now?"
-    )
-
-    result = messagebox.askyesno("Accessibility Permission", message)
-    if result:
-        try:
-            import subprocess
-
-            subprocess.Popen(["open", "x-apple.systempreferences:com.apple.preference.security?Privacy_Accessibility"])
-        except Exception as e:
-            logging.error(f"Failed to open System Preferences: {e}")
-
-        # Accessibility (TCC) changes only take effect on relaunch.
-        messagebox.showinfo(
-            "Restart Required",
-            "Accessibility permission only takes effect after Edgeware++ is restarted.\n\n"
-            "Edgeware++ will now close.  Please reopen it once you have granted permission.",
-        )
-        root.after(100, root.destroy)
-
-
-def check_input_monitoring_permission() -> bool:
-    """Check if the app has Input Monitoring permission on macOS.
-
-    AXIsProcessTrusted() only proves Accessibility (which CGEventTap creation does
-    NOT require).  The reliable signal for a global keyboard event tap is whether
-    CGEventTapCreate itself returns a valid tap: if Input Monitoring has not been
-    granted to this app, it returns None immediately.
-
-    Returns True if permission is granted or if not on macOS.
-    Returns False if permission is not granted or if check fails.
-    """
-    if sys.platform != "darwin":
-        return True
-
-    try:
-        import Quartz
-
-        # Probe with a minimal listen-only keyboard CGEventTap.  We never enable
-        # it or attach it to a run loop; its creation is the permission signal.
-        tap = Quartz.CGEventTapCreate(
-            Quartz.kCGSessionEventTap,
-            Quartz.kCGHeadInsertEventTap,
-            Quartz.kCGEventTapOptionListenOnly,
-            Quartz.CGEventMaskBit(Quartz.kCGEventKeyDown),
-            lambda proxy, event_type, event, refcon: event,
-            None,
-        )
-        granted = tap is not None
-        logging.info(f"Input Monitoring permission check: {'granted' if granted else 'not granted'}")
-        return granted
-    except Exception as e:
-        logging.warning(f"Could not check Input Monitoring permission: {e}")
-        return False
-
-
-def show_input_monitoring_dialog(root: Tk) -> None:
-    """Show dialog explaining the Input Monitoring permission requirement.
-
-    If the user opts to open System Settings, warn that a restart is required for
-    the change to take effect, then quit the app so they can relaunch it.  If
-    they decline, keep running (the tray / IPC panic path still works).
-
-    The correct macOS pane for global keyboard event taps is the Input Monitoring
-    list under Privacy & Security, not Accessibility.
-    """
-    from tkinter import messagebox
-
-    message = (
-        "Edgeware++ needs Input Monitoring permission to listen for the global panic hotkey.\n\n"
-        "To grant permission:\n"
-        "1. Open System Settings\n"
-        "2. Go to Privacy & Security > Input Monitoring\n"
-        "3. Enable Edgeware++\n\n"
-        "If you don't want to grant this permission, you can still use:\n"
-        "• The Panic.app (separate application)\n"
-        "• The tray icon menu\n\n"
-        "Do you want to open System Settings now?"
-    )
-
-    result = messagebox.askyesno("Input Monitoring Permission", message)
-    if result:
-        try:
-            import subprocess
-
-            subprocess.Popen(
-                ["open", "x-apple.systempreferences:com.apple.preference.security?Privacy_ListenEvent"]
-            )
-        except Exception as e:
-            logging.error(f"Failed to open System Settings: {e}")
-
-        # Input Monitoring (TCC) changes only take effect on relaunch.
-        messagebox.showinfo(
-            "Restart Required",
-            "Input Monitoring permission only takes effect after Edgeware++ is restarted.\n\n"
-            "Edgeware++ will now close.  Please reopen it once you have granted permission.",
-        )
-        root.after(100, root.destroy)
 
 
 def open_web(pack: Pack, web: str | None = None) -> None:
@@ -429,7 +277,7 @@ def _handle_keyboard_darwin(root: Tk, settings: Settings, state: State) -> None:
     in-process on a daemon thread and is polled via Tk's after(), avoiding both
     subprocess permission inheritance and pynput's macOS TSM main-thread crash.
     """
-    from features.keyboard_listener_darwin import DarwinKeyboardListener
+    from os_utils.keyboard_listener_darwin import DarwinKeyboardListener
 
     def on_panic() -> None:
         if state.panic_shutdown:
